update recipe description

+++ controllers/recipeController.js
@@ -5,15 +5,16 @@ const recipeModel = require("../models/recipeModel")
 const commonHelper = require("../helper/common")
 const uuid = require("uuid")
 var cloudinary = require("../config/cloudinary")
+const moment = require("moment")
 
 const recipeController = {
   getAllRecipe: async (req, res) => {
     try {
       const page = Number(req.query.page) || 1
-      const limit = Number(req.query.limit) || 10000
+      const limit = Number(req.query.limit) || 5
       const offset = (page - 1) * limit
-      const sortBY = req.query.sortBY || "name_recipe"
-      const sort = req.query.sort || "ASC"
+      const sortBY = req.query.sortBY || "created_at"
+      const sort = req.query.sort || "desc"
       const searchParam = req.query.search ? req.query.search.toLowerCase() : ""
       const result = await recipeModel.selectAllRecipe(limit, offset, searchParam, sortBY, sort)
       if (result.rowCount === 0) {
@@ -32,9 +33,20 @@ const recipeController = {
         totalData,
         totalPage,
       }
-      commonHelper.response(res, result.rows, 200, "get data success", pagination)
-    } catch (error) {
-      console.log(error)
+      commonHelper.response(
+        res,
+        result.rows.map((data) => {
+          return {
+            ...data,
+            created_at: moment(data.created_at).format("DD MMMM YYYY HH:mm"),
+          }
+        }),
+        200,
+        "Get data success",
+        pagination
+      )
+    } catch (err) {
+      res.json({ message: err.message })
     }
   },
 
@@ -49,41 +61,46 @@ const recipeController = {
     recipeModel
       .selectRecipe(id)
       .then((result) => {
-        commonHelper.response(res, result.rows[0], 200, "Get data success")
+        const data = {
+          ...result.rows[0],
+          created_at: moment(result.rows[0].created_at).format("DD MMMM YYYY HH:mm")
+        }
+        commonHelper.response(res, data, 200, "Get data success")
       })
       .catch((err) => res.send(err))
   },
 
   createRecipe: async (req, res) => {
-    const { name_recipe, ingredients, video } = req.body
+    const { name_recipe, description, ingredients, video } = req.body
     const id = uuid.v4()
-  
+
     const imageUrl = await cloudinary.uploader.upload(req.file.path, { folder: "izipizy" })
-  
+
     // Extract the user ID from the decoded token
     const userId = req.payload.id
-  
+
     const data = {
       id,
       name_recipe,
+      description,
       ingredients,
       video,
       image: imageUrl.secure_url,
       user_id: userId,
     }
-  
+
     recipeModel
       .insertRecipe(data)
       .then((result) => {
         commonHelper.response(res, result.rows, 201, "Recipe has been created")
       })
       .catch((err) => res.send(err))
-  },  
+  },
 
   updateRecipe: async (req, res) => {
     try {
       const id = req.params.id
-      const { name_recipe, ingredients, video } = req.body
+      const { name_recipe, description, ingredients, video } = req.body
       // Check if the recipe exists and if the user who created it is the same as the authenticated user
       const { rows } = await recipeModel.findId(id)
       if (rows.length === 0) {
@@ -91,13 +108,25 @@ const recipeController = {
           message: "Recipe not found",
         })
       }
+
+      // Extract the user ID from the decoded token
+      const userId = req.payload.id
+
+      // Check if the user ID of the decoded token matches the user ID of the recipe
+      if (rows[0].user_id !== userId) {
+        return commonHelper.response(res, null, 401, "You are not authorized to edit this recipe")
+      }
+
       let data = {}
       let updateQuery = ""
       let message = "Recipe updated sucessfull"
-
       if (name_recipe) {
         data.name_recipe = name_recipe
         updateQuery += `name_recipe=$${Object.keys(data).length}`
+      }
+      if (description) {
+        data.description = description
+        updateQuery += `${updateQuery ? ", " : ""}description=$${Object.keys(data).length}`
       }
       if (ingredients) {
         data.ingredients = ingredients
@@ -133,12 +162,19 @@ const recipeController = {
   deleteRecipe: async (req, res) => {
     try {
       const id = req.params.id
-      const { rowCount } = await recipeModel.findId(id)
+      const { rowCount, rows } = await recipeModel.findId(id)
       if (!rowCount) {
-        return res.json({
-          Message: "Recipe not found",
-        })
+        return commonHelper.response(res, null, 404, "Recipe not found")
       }
+
+      // Extract the user ID from the decoded token
+      const userId = req.payload.id
+
+      // Check if the user ID of the decoded token matches the user ID of the recipe
+      if (rows[0].user_id !== userId) {
+        return commonHelper.response(res, null, 401, "You are not authorized to delete this recipe")
+      }
+
       recipeModel
         .deleteRecipe(id)
         .then((result) => {
